backend/pyproject.toml
[tool.poetry]
name = "llmany-backend"
version = "0.1.0"
description = ""
authors = ["Oskar Kulinski", "Kajetan Ożóg", "Michał Smorągiewicz"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.57.1"
google-generativeai = "^0.8.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
pytest = "^8.3.3"
pytest-mock = "^3.14.0"
mypy = "^1.13.0"

[tool.pytest.ini_options]
addopts = [
    "--import-mode=importlib",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
